Fix some type mismatch

+++ final/GameofLife/app/src/main/java/mobileappdevelopment/kevinholmes/gameoflife/DatabaseHelper.java
@@ -80,7 +80,7 @@ public class DatabaseHelper extends SQLiteOpenHelper{
     }
 
     //Allows
-    public SerializableCellGrid requestGrid(int id){
+    public SerializableCellGrid requestGrid(Long id){
         // Execute SQL to retrieve thing with proper name
         // the new byte array will be replaced by actual data once this is working
